allow transparent output from various commands (e.g., duplicity, apt-get)
Background:

* we occasionally need to capture output for logging purposes, to
  check for errors based on output (e.g., we look at duplicity's
  outputs on error to guess if it failed because of a bad address, or
  something else such as a bad key), etc.

* formerly we used commands.getstatusoutput to capture output

* the problem with these commands is that they don't let you see
  partial output - you have to wait until the command is finished.
  Duplicity will fail very slowly in some conditions (e.g., SSH key
  is unknown). apt-get update and apt-get install can also take
  time.  If you're running restore non-interactively it doesn't
  matter, but otherwise it is helpful, especially when things don't
  seem to be working right, to get immediate feedback on what is
  happening.

Solution:

* Use stdtrap to intercept output to stdout transparently if we're
  running the command interactively (e.g., we can tell if stdout is a
  tty, or just a pipe to a file or command)

+++ utils.py
@@ -1,6 +1,11 @@
+import sys
 import os
 from os.path import *
+
+import tempfile
 import shutil
+
+from stdtrap import UnitedStdTrap
 
 def remove_any(path):
     """Remove a path whether it is a file or a directory. 
@@ -19,8 +24,40 @@ def remove_any(path):
 class TempDir(str):
     def __new__(cls, prefix='tmp', suffix='', dir=None):
         path = tempfile.mkdtemp(suffix, prefix, dir)
-        return str.__new__(cls, path)
+        self = str.__new__(cls, path)
+        self.pid = os.getpid()
+
+        return self
+
+    def remove(self):
+        if exists(self):
+            shutil.rmtree(self)
 
     def __del__(self):
-        shutil.rmtree(self)
+        if self.pid == os.getpid():
+            self.remove()
+
+def system(command, transparent=None):
+    """Run a command and intercept its output 
+       Returns: (status, output)
+
+    Arguments:
+    <command>       Command to pass to os.system()
+    <transparent>   Allow output to reach stdout "transparently"?
+                    By default: True if stdout is a tty, False otherwise
+    """
+    if transparent is None:
+        if os.isatty(sys.stdout.fileno()):
+            transparent = True
+        else:
+            transparent = False
+
+    trap = UnitedStdTrap(transparent=transparent)
+    try:
+        status = os.system(command)
+    finally:
+        trap.close()
+    output = trap.std.read()
+
+    return status, output
